+++ src/main/java/com/zufar/icedlatte/user/api/avatar/UserAvatarLinkProvider.java
@@ -1,6 +1,6 @@
 package com.zufar.icedlatte.user.api.avatar;
 
-import com.zufar.icedlatte.common.filestorage.api.FileProvider;
+import com.zufar.icedlatte.filestorage.file.FileProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
