+++ src/main/java/com/zufar/icedlatte/user/api/avatar/UserAvatarUploader.java
@@ -1,8 +1,9 @@
 package com.zufar.icedlatte.user.api.avatar;
 
-import com.zufar.icedlatte.common.filestorage.api.FileUploader;
-import com.zufar.icedlatte.common.filestorage.api.MinioFileService;
-import com.zufar.icedlatte.common.filestorage.dto.FileMetadataDto;
+import com.zufar.icedlatte.filestorage.filemetadata.FileMetadataDeleter;
+import com.zufar.icedlatte.filestorage.filemetadata.FileMetadataSaver;
+import com.zufar.icedlatte.filestorage.file.FileUploader;
+import com.zufar.icedlatte.filestorage.dto.FileMetadataDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -21,15 +22,16 @@ public class UserAvatarUploader {
     private String bucketName;
     private static final String AVATAR_NAME_PREFIX = "user-avatar-";
 
-    private final MinioFileService minioFileService;
     private final FileUploader fileUploader;
+    private final FileMetadataSaver fileMetadataSaver;
+    private final FileMetadataDeleter fileMetadataDeleter;
 
     @Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.READ_COMMITTED)
     public FileMetadataDto uploadUserAvatar(final UUID userId, final MultipartFile file) {
-        minioFileService.deleteByRelatedObjectId(userId);
+        fileMetadataDeleter.deleteByRelatedObjectId(userId);
         String fileName = AVATAR_NAME_PREFIX + userId.toString();
         fileUploader.upload(file, bucketName, fileName);
         FileMetadataDto fileMetadataDto = new FileMetadataDto(userId, bucketName, fileName);
-        return minioFileService.save(fileMetadataDto);
+        return fileMetadataSaver.save(fileMetadataDto);
     }
 }
